requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Form handling
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
aiosqlite>=0.19.0

# Development (optional)
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.7.0
